moved chain prio methods to no executor
