+++ rclcpp/src/rclcpp/executor.cpp
@@ -262,6 +262,7 @@ Executor::add_node(rclcpp::node_interfaces::NodeBaseInterface::SharedPtr node_pt
             std::string("Node '") + node_ptr->get_fully_qualified_name() +
             "' has already been added to an executor.");
   }
+
   {
     std::lock_guard<std::mutex> guard{mutex_};
     node_ptr->for_each_callback_group(
@@ -280,8 +281,6 @@ Executor::add_node(rclcpp::node_interfaces::NodeBaseInterface::SharedPtr node_pt
 
     weak_nodes_.push_back(node_ptr);
   }
-
-  apply_chain_priorities();
 }
 
 void
@@ -345,133 +344,6 @@ Executor::add_node(std::shared_ptr<rclcpp::Node> node_ptr, bool notify)
   this->add_node(node_ptr->get_node_base_interface(), notify);
 }
 
-void
-Executor::set_chain_priority_allocator(
-  std::shared_ptr<rclcpp::detail::ChainPriorityAllocator> allocator)
-{
-  std::lock_guard<std::mutex> guard{mutex_};
-  chain_priority_allocator_ = std::move(allocator);
-}
-
-void
-Executor::apply_chain_priorities()
-{
-  std::lock_guard<std::mutex> guard{mutex_};
-  if (!chain_priority_allocator_) {
-    return;
-  }
-
-  std::unordered_map<std::string, rclcpp::CallbackGroup::SharedPtr> groups_by_name;
-  std::unordered_map<std::string, std::shared_ptr<rclcpp::sched::SchedBase>> entities_by_name;
-
-  const auto logger = rclcpp::get_logger("Executor");
-  auto register_named_entity =
-    [&groups_by_name, &entities_by_name, logger](
-    const std::string & name,
-    const rclcpp::CallbackGroup::SharedPtr & callback_group,
-    const std::shared_ptr<rclcpp::sched::SchedBase> & entity)
-    {
-      if (name.empty() || !callback_group || !entity) {
-        return;
-      }
-      auto [entity_it, inserted] = entities_by_name.emplace(name, entity);
-      if (!inserted) {
-        RCLCPP_WARN(
-          logger,
-          "Duplicate callback name '%s'; keeping first registration",
-          name.c_str());
-        return;
-      }
-      groups_by_name.emplace(name, callback_group);
-    };
-
-  for (const auto & pair : weak_groups_to_nodes_) {
-    auto group = pair.first.lock();
-    if (!group) {
-      continue;
-    }
-    group->collect_all_ptrs(
-      [&register_named_entity, &group](const rclcpp::SubscriptionBase::SharedPtr & subscription) {
-        if (subscription) {
-          register_named_entity(
-            subscription->get_callback_name(),
-            group,
-            std::static_pointer_cast<rclcpp::sched::SchedBase>(subscription));
-        }
-      },
-      [&register_named_entity, &group](const rclcpp::ServiceBase::SharedPtr & service) {
-        if (service) {
-          register_named_entity(
-            service->get_callback_name(),
-            group,
-            std::static_pointer_cast<rclcpp::sched::SchedBase>(service));
-        }
-      },
-      [&register_named_entity, &group](const rclcpp::ClientBase::SharedPtr & client) {
-        if (client) {
-          register_named_entity(
-            client->get_callback_name(),
-            group,
-            std::static_pointer_cast<rclcpp::sched::SchedBase>(client));
-        }
-      },
-      [&register_named_entity, &group](const rclcpp::TimerBase::SharedPtr & timer) {
-        if (timer) {
-          register_named_entity(
-            timer->get_callback_name(),
-            group,
-            std::static_pointer_cast<rclcpp::sched::SchedBase>(timer));
-        }
-      },
-      [&register_named_entity, &group](const rclcpp::Waitable::SharedPtr & waitable) {
-        if (waitable) {
-          register_named_entity(
-            waitable->get_callback_name(),
-            group,
-            std::static_pointer_cast<rclcpp::sched::SchedBase>(waitable));
-        }
-      });
-  }
-
-  if (groups_by_name.empty()) {
-    RCLCPP_WARN(
-      logger,
-      "Chain priority allocator is set but no named callbacks are registered");
-    return;
-  }
-
-  auto allocation = chain_priority_allocator_->allocate(groups_by_name);
-  for (const auto & pair : allocation.callback_priorities) {
-    const auto & callback_name = pair.first;
-    const auto priority = pair.second;
-    auto entity_it = entities_by_name.find(callback_name);
-    if (entity_it == entities_by_name.end()) {
-      RCLCPP_WARN(
-        logger,
-        "No callback entity registered for '%s'; skipping priority assignment",
-        callback_name.c_str());
-      continue;
-    }
-    auto & entity = entity_it->second;
-    if (!entity) {
-      RCLCPP_WARN(
-        logger,
-        "Callback entity for '%s' is no longer valid; skipping priority assignment",
-        callback_name.c_str());
-      continue;
-    }
-
-    entity->set_edf_attr(nullptr);
-    rclcpp::sched::SchedAttr attr = entity->sched_attr;
-    attr.sched_policy = SCHED_FIFO;
-    attr.sched_priority = priority;
-    attr.sched_flags = 0;
-    attr.sched_runtime = 0;
-    attr.sched_deadline = 0;
-    attr.sched_period = 0;
-    entity->set_sched_attr(attr);
-  }
-}
 
 void
 Executor::remove_node(rclcpp::node_interfaces::NodeBaseInterface::SharedPtr node_ptr, bool notify)
