+++ rclcpp/src/rclcpp/executors/single_threaded_executor.cpp
@@ -724,4 +724,3 @@ SingleThreadedExecutor::spin_forever()
     // num_cb_dispatched++;
   }
 }
-
